services/esp32_manager.py
#!/usr/bin/env python3
"""
Manager robusto para comunicación con ESP32
ARCHIVO PRINCIPAL - Ver implementación completa en la documentación
"""

import serial
import json
import time
import threading
import asyncio
from datetime import datetime, timedelta
from typing import Optional, Dict, Any, Callable
import queue
from core.config import settings
from core.logger import logger
from models.esp32_data import ESP32Data

class ESP32CommunicationError(Exception):
    """Excepción para errores de comunicación"""
    pass

class ESP32Manager:
    """Manager principal para comunicación con ESP32"""
    
    def __init__(self):
        self.serial_conn: Optional[serial.Serial] = None
        self.connected = False
        self.last_command_time = 0
        self.lock = threading.Lock()
        self.running = False
        
        # Cache y estado
        self.last_data: Optional[Dict[str, Any]] = None
        self.last_successful_communication = time.time()
        self.communication_errors = 0
        
        # Cola de comandos para rate limiting
        self.command_queue = queue.Queue(maxsize=10)
        self.worker_thread: Optional[threading.Thread] = None
        
        # Rate limiting
        self.request_times = []
        
    async def start(self) -> bool:
        """Inicializar manager y conectar"""
        logger.info("🚀 Iniciando ESP32 Manager...")
        
        if await self._connect():
            self.running = True
            self._start_worker_thread()
            logger.info("✅ ESP32 Manager iniciado correctamente")
            return True
        
        logger.error("❌ Error iniciando ESP32 Manager")
        return False
    
    async def _connect(self) -> bool:
        """Establecer conexión serial"""
        max_retries = 3
        
        for attempt in range(max_retries):
            try:
                if self.serial_conn and self.serial_conn.is_open:
                    self.serial_conn.close()
                
                self.serial_conn = serial.Serial(
                    port=settings.SERIAL_PORT,
                    baudrate=settings.SERIAL_BAUDRATE,
                    timeout=settings.SERIAL_TIMEOUT,
                    bytesize=serial.EIGHTBITS,
                    parity=serial.PARITY_NONE,
                    stopbits=serial.STOPBITS_ONE
                )
                
                # Limpiar buffers
                self.serial_conn.reset_input_buffer()
                self.serial_conn.reset_output_buffer()
                
                # Probar comunicación
                if await self._test_communication():
                    self.connected = True
                    self.communication_errors = 0
                    logger.info(f"✅ Conectado al ESP32 en {settings.SERIAL_PORT}")
                    return True
                
            except Exception as e:
                logger.warning(f"❌ Intento {attempt + 1}: {e}")
                await asyncio.sleep(1)
        
        self.connected = False
        return False
    
    async def get_data(self) -> Optional[ESP32Data]:
        """Obtener datos completos del ESP32"""
        # Implementación simplificada para demo
        logger.info("📊 Solicitando datos del ESP32...")
        
        # Aquí iría la implementación completa del protocolo serial
        # Por ahora retorna datos de ejemplo
        try:
            # Datos de ejemplo para pruebas
            sample_data = {
                "panelToBatteryCurrent": 2450.0,
                "batteryToLoadCurrent": 1850.2,
                "voltagePanel": 18.45,
                "voltageBatterySensor2": 13.28,
                "currentPWM": 145,
                "temperature": 42.3,
                "chargeState": "ABSORPTION_CHARGE",
                "bulkVoltage": 14.4,
                "absorptionVoltage": 14.4,
                "floatVoltage": 13.6,
                "LVD": 12.0,
                "LVR": 12.5,
                "batteryCapacity": 100.0,
                "thresholdPercentage": 2.5,
                "maxAllowedCurrent": 8000.0,
                "isLithium": False,
                "maxBatteryVoltageAllowed": 15.0,
                "absorptionCurrentThreshold_mA": 2500.0,
                "currentLimitIntoFloatStage": 500.0,
                "calculatedAbsorptionHours": 1.25,
                "accumulatedAh": 67.8,
                "estimatedSOC": 68.5,
                "netCurrent": 599.8,
                "factorDivider": 5,
                "useFuenteDC": False,
                "fuenteDC_Amps": 0.0,
                "maxBulkHours": 0.0,
                "panelSensorAvailable": True,
                "maxAbsorptionHours": 1.0,
                "chargedBatteryRestVoltage": 12.88,
                "reEnterBulkVoltage": 12.6,
                "pwmFrequency": 40000,
                "tempThreshold": 55,
                "temporaryLoadOff": False,
                "loadOffRemainingSeconds": 0,
                "loadOffDuration": 0,
                "loadControlState": True,
                "ledSolarState": True,
                "notaPersonalizada": "Sistema funcionando correctamente",
                "connected": True,
                "firmware_version": "ESP32_v2.1",
                "uptime": 847293,
                "last_update": str(int(time.time()))
            }
            
            return ESP32Data(**sample_data)
            
        except Exception as e:
            logger.error(f"❌ Error obteniendo datos: {e}")
            return None
    
    async def set_parameter(self, parameter: str, value: Any) -> bool:
        """Establecer un parámetro"""
        logger.info(f"⚙️ Configurando {parameter} = {value}")
        
        # Aquí iría la implementación real del protocolo serial
        # Por ahora simula éxito
        await asyncio.sleep(0.1)  # Simular delay de comunicación
        return True
    
    async def toggle_load(self, total_seconds: int) -> bool:
        """Apagar carga temporalmente"""
        logger.info(f"🔌 Apagando carga por {total_seconds} segundos")
        
        # Aquí iría la implementación real
        await asyncio.sleep(0.1)
        return True
    
    async def cancel_temporary_off(self) -> bool:
        """Cancelar apagado temporal"""
        logger.info("🔌 Cancelando apagado temporal")
        
        # Aquí iría la implementación real
        await asyncio.sleep(0.1)
        return True
    
    def get_connection_info(self) -> Dict[str, Any]:
        """Obtener información de conexión"""
        return {
            "connected": self.connected,
            "port": settings.SERIAL_PORT,
            "baudrate": settings.SERIAL_BAUDRATE,
            "last_communication": self.last_successful_communication,
            "communication_errors": self.communication_errors,
            "queue_size": 0
        }
    
    async def stop(self):
        """Detener manager"""
        logger.info("🛑 Deteniendo ESP32 Manager...")
        self.running = False
        if self.serial_conn and self.serial_conn.is_open:
            self.serial_conn.close()
        self.connected = False

# NOTA: Esta es una versión simplificada para el setup inicial.
# Para la implementación completa, consulta la documentación del proyecto.
