requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyserial==3.5
redis==5.0.1
python-multipart==0.0.6
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
